+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## New since 0.20.0
 
+- Canvas adapter: line scale option [#1075](https://github.com/mapsforge/vtm/pull/1075)
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.21.0)